pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "file diff"
authors = ["Nevelskoy <safelikc@gmail.com>"]
readme = "README.md"
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiffer = "gendiff.scripts.gendiff:main"
